pyproject.toml
[project]
name = "mcp-bildsprache"
version = "0.3.0"
description = "MCP server for brand-aware image generation — Gemini, FLUX.2 Pro, Recraft V3"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [{ name = "Casey Romkes" }]
dependencies = [
    "fastmcp>=3.2.2",
    "httpx>=0.28.0",
    "pydantic>=2.10",
    "pydantic-settings>=2.12.0",
    "pillow>=11.0.0",
    "piexif>=1.1.3",
    "python-slugify>=8.0.0",
]

[project.scripts]
mcp-bildsprache = "mcp_bildsprache.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "pytest-httpx>=0.34",
    "ruff>=0.8",
]
